Creating a data handler base class
